Add glide.yaml for vendor packages
Until such time as the fuchsia build system has a two-staged build,
there needs to be a way to manage vendor packages.  The Go team has
conveniently provided a solution for this via the vendor/ directory,
which acts as a store for the third-party packages required by a
particular package without polluting the global GOPATH.

Glide (https://github.com/Masterminds/glide) is a tool for managing
the packages inside the vendor/ directory.  It looks for a single
glide.yaml file in the package directory that lists all the dependencies
for that package as well as the specific versions of those
dependencies (if any).

Add a glide.yaml file containing the only dependency of thinfs so
far (glog).

Change-Id: I0f5f4301572b58621f35d1002ee9335cb7ce2c7e
Signed-off-by: Chirantan Ekbote <chirantan@google.com>