Add a libext2fs io_manager for fuchsia systems
Add a full fuchsia-specific implementation of the io_manager interface
that libext2fs expects for doing I/O operations.  Building package
ext2fs with the fuchsia io_manager requires building with the 'fuchsia'
build tag.

Also clean up some issues with the runtime.SetFinalizer call in
ext2fs.New().  The function in the second parameter should accept a *FS
as a parameter and the finalizer should be set up as soon as the
ext2_filsys object has been created.

Change-Id: Ife4a713b6d47b6f5145b371ece122bc2851413f2
Signed-off-by: Chirantan Ekbote <chirantan@google.com>