Add mojom interfaces that wrap the file system service
Add mojoms that define the interfaces between the file system service
and its clients as well as between the file system service and the block
device service.

The file system interface has been kept as simple as possible while
still supporting the needs of syncbase.  More operations can easily be
added to the interface in the future once we have clients that actually
need them.

Eventually these interfaces should probably go into a common repository
with the interfaces for all the other fuchsia system services but we can
keep them here for now until we actually have some of the other services
running.

Change-Id: Ia5046bae5f36dfa4c93f4815baf35fe3688780f5
Signed-off-by: Chirantan Ekbote <chirantan@google.com>