Add dependency on github.com/pkg/errors
The errors package provides a nice way to annotate errors with context
about where and how the error occurred.  See
http://dave.cheney.net/2016/04/27/dont-just-check-errors-handle-them-gracefully
for more details on the package.

Also exclude all the packages in the mojo/ directory for now.  glide
gets really confused about packages that come from a different directory
in the GOPATH.

Change-Id: I9acaa8b8de5d3d8479938647714d5e954f8741f1
Signed-off-by: Chirantan Ekbote <chirantan@google.com>