[mp] Refactor mp_sync_exec for upcoming CPU hotplug support
This reworks the internal accounting of which CPUs need to perform
the task to be in terms of a CPU mask rather than a count.  This
allows us to detect that a CPU was unplugged and to stop waiting
for it to execute the task.

Also fix an inconsistency between SMP/non-SMP mp_sync_exec.  On
non-SMP, the task was being executed potentially with interrupts
enabled.

Change-Id: I7c6229f81cc2890558627f3e4b60002e76a938c9