[system][mxio] Discard handles during remoteio handler error.
Currently, when the remoteio handler receives a malformed message, it
sets msg.arg to "ERR_IO", but only discards the handles when the
"msg.hcount" field is malformed. If the "msg.datalen" field is
malformed, then the handles are still sent, along with the "ERR_IO".

If mx_rio_txn sees an error in the "msg.arg" value, it discards all
handles anyway. In the case an error occurs, why not short-circuit and
prevent those handles from being sent in the first place?

Change-Id: I4c1bb9c56dce84e9af40cba701dda7d36a7b30f4