[unittest] Reduce duplication in the EXPECT_* macros
The reason for doing this is that I would like to add some fail-fast
ASSERT_*() macros (which would abort rather than continuing if the
comparison is false), and I would like to avoid duplicating more code
when doing that.

Note that this can change what warnings are generated for some uses of
EXPECT_NEQ().

This changes the output text a bit, especially for EXPECT_NEQ().

Change-Id: I2acb840d8f36512198217ba792b6a375e75c52ae