[system][mxio] Changed remoteio 'off' field into 'arg2' union.
The 'off' field of remoteio had been used as a secondary argument
already -- sometimes it was used as an offset, sometimes as something
else. It's confusing to see something like "msg->off = PROTOCOL_FOO" in
code, since it discards the intuition behind the word 'offset'.

Instead, these usages of the secondary argument in the mx_rio_msg have
been replaced with a union, which will allow more appropriate naming /
type conventions for aruments passed in the mx_rio_msg header.

Change-Id: I2aa55c6263d9fcec1c9ef29a63153855e8783aac