[ulib][ddk] driver binding/matching "language"
A very simple virtual machine for matching drivers against devices.

The idea here is instead of having to call a probe() hook or even
load the driver, we can "execute" a simple little program that will
evaluate whether the device in question is supportable by the driver.

The driver can make more detailed decisions in its init() hook if
necessary, but this should eliminate the need for boiler-plate in
probe().

Change-Id: Iffd1d2e814f56291ec96ed4e73ed86ad81a1838b