Renamed "handle" package to "cpointer".
The purpose of the handle package is to pair "Go values" with arbitrary
"C pointers", so the Go values can be referred to by keys that can pass
across language barriers. Unfortunately, "handle" is an overloaded term
in the context of Mojo and Magenta, and "cpointer" is a more appropriate
name.

Change-Id: Id3bb15b9a86eca607545f1f9bed7bb2d671c03b2