[musl] disable the weak symbol aliased io plumbing until we have solibs
When linking .a libraries, the weak aliasing trick for letting mxio
override the io plumbing does not work, so unless WITH_LIBC_SO is
defined, we simply make __lib_io_write() and friends external instead
of weakly aliased.

This does mean if you want to link with libc but not libmxio you need
to provide those functions, but currently we don't do that anywhere
in the build.

Once we are building and using shared libraries we can return to the
state where libc has no-op io plumbling by default, overrideable at
link time.

Change-Id: I0e7511126511a2e5c7a82ae769cac096007674b6