[ulib] [mxio] Add auxiliary vector to mx_proc_args_t
This extends mx_proc_args_t to hold the "auxiliary vector", an
array of pairs of integers that is the traditional Unixoid way to
pass extra startup details from the system ELF loader to the
dynamic linker and on to the main executable.  In a future change
the mxio process-starting code will pass essential information
from the ELF loader to the new process this way.

Change-Id: I6becd939b2dc9880660e54e01dedacf4d1e7d591