[message-pipe] Test error cases on reading from closed/empty pipes
This tests reading from message pipes that are empty/nonempty with
the peer handle closed/not closed. The test passes, but this reveals
that we're folding the error codes for no-message-but-peer-not-closed
and no-message-and-peer-closed into the same value, which is not very
useful as callers will want to distinguish these two cases and in mojo
these are distinct values.

Change-Id: Id945a9b6973075d460f968c8ae2d0b564655cbd7