[pcie] Add support for PCIe caps v1
Add support for parsing the PCI Express Capabilities structure version
1.  Unlike v2 structures, where the spec seems to state that all
registers will always be present (although some will be hardwired to
zero), the 1.0, 1.1 and 2.0 PCIe specs appear to imply that registers
may be totally absent (not just hardwired to zero) depending on the
specific device type.

So... Update the code to pay attention to the version number and to
figure out the variable size of the v1 structure based on device type.

Change-Id: If2b7382be03b35ffabe6913560fdaced6e0385a7