[mxio] Support ET_DYN, PT_INTERP in userland ELF loader
This extends the mxio ELF loader to handle ELF files with e_type
ET_DYN (DSOs and PIEs, which get loaded at an address chosen by
the kernel), and to handle the PT_INTERP program header, which
points to a second ELF file (by name) that will be loaded in the
same address space.  PT_INTERP is the usual method by which a
dynamic linker gets loaded and has its entry point run (so it is
responsible for eventually jumping to the main executable's entry
point).

The new elf-interp test is a minimal test of the ELF loader's
PT_INTERP handling.  It builds a tiny helper DSO that gets loaded
via PT_INTERP.  The helper has the minimal code that can be used
to verify PT_INTERP loading works, without having to write a
significant portion of a dynamic linker's code for the test.
Keeping the code minimal requires building it in a rather kludgy
fashion where it hard-codes addresses in the main executable image.

Change-Id: I1f051881732e907beb16c13605324cf23510528d