Added and tested (poorly performing) thread safety to thinio conductor
Previously, concurrent access to the conductor could cause race
conditions in the cache. As a result, interfaces expecting to issue
multiple requests to the device concurrently were unable to do so.

Now, higher-level interfaces (such as filesystems) can issue multiple
requests to the conductor at the same time. Unfortunately, under the
hood, the conductor serializes access to the cache and device. A more
finely-grained implementation will be needed for reasonable performance,
but the interface to the conductor should stay the same.

Change-Id: I276d5bfb66cf8ca66f7099dc2533b726dff53df2