[build] build all ulib, uapp, and utest modules in system and third_party
Which is what we do now, but let's not require people to muck with
kernel/project/virtual/user.mk every time a module is added or removed.

At some point we may not want to include every module in every final
image, but that should be a separate configuration / selection step.

Change-Id: I183ab1f5d11a924fba0e30659befef79bdd06f08