[magenta] some header cleanup
Move the syscall-parameter-specific types to syscalls-types.h,
leaving magenta/types.h just the more general types and error
codes.

We may sub-categorize this later, but the upshot is we don't
bring very specific types and datastructures into scope for
things that just want to use common error codes or typedefs.

Change-Id: I1ca83db36994e0e59de925db7567d1867f63fa7a