Build mkbootfs in platform-specific path
This builds mkbootfs into linux64/mkbootfs or mac/mkbootfs
depending on the host platform and makes the root mkbootfs a
shell script that runs the right binary.

Also teaches update.sh to rebuild mkbootfs if the actual binary
is missing, instead of only if the stamp file is missing or incorrect.

Change-Id: I9420dac3337c838e4ad9a1d8f088d394ceed481a