[unittest] Implement REQUIRE_* forms of EXPECT_*
Implement REQUIRE_ versions of the EXPECT_ macros which will terminate
a test immdiately if the test condition is not satisfied.

Also, do a little more cleanup on the EXPECT_BYTES_EQ and NE macros.
Make sure that when they hexdump, they go through the unittest print
hooks, so their output ends up in the proper output stream.

Change-Id: Ic604f29c2541dabd1dcb1380f3cb5dd0827d8a8a