[magenta] Replace use of sys/types.h with stdint.h
sys/types.h is a POSIX header files and is not a part of C
standard. There are number of defines in sys/types.h which are
guarded by either _GNU_SOURCE or _BSD_SOURCE. We should be using
stdint.h and stddef.h instead which are C standard equivalents.

Change-Id: Idd9642504140a59079f07ab8e487f94b1f5b4cf5