[magenta][syscalls] Refactor wait-one/wait-many to use a WaitEvent object.
Also make wait-one/wait-many provide the satisfied/satisfiable signals
in a struct (mx_signals_state_t), and add a result_index output argument
to wait-many.

Currently, the WaitEvent object still uses an event_t, but that can be
changed later.

The WaitEvent can pass a context, which is used to transmit the
result_index for wait-many (i.e., the index of the first handle to
cause the wait to be satisfied).

Some of the abstraction (like WaitEvent::Result) is to support future
work on wait sets, etc. (The context also is a uint64_t, to support work
on wait sets.)

Change-Id: I54c517b8627d225da72a23ff5a72be9b3a92065b