[musl] Move "heap donation" magic into malloc implementation
The dynamic linker's "reclaim" function is full of malloc implementation
details.  Factor those out into a subroutine that lives in the malloc
implementation, so it's more clear what the magic is doing.  This way,
if the malloc implementation is changed or replaced then the dynamic
linker code won't need arcane adjustments.

Change-Id: Ia574b10fc2c9a1c55b003948f9a6640f8c772263