[musl] Copy writable segments to simulate copy-on-write
The kernel does not yet support copy-on-write mapping semantics.
The dynamic linker needs those semantics for writable segments.
Otherwise it modifies the VM object representing the file, which
is also used to back the read-only segments and in canonical ELF
layout the read-only and writable portions will overlap in the file,
so touching the writable data also touches some read-only pages.
To work around the problem, eagerly copy data into a fresh VM
object for a writable segment.

Change-Id: Ib3b762613b8411c61644c794397458f32faf808b