[utest] Clean up more minutiae in utests
This moves all the remaining printfs in unit tests (but not other tests,
or helper programs) to unittest_printf.

This also moves thread-local-storage to pthread-tls to make it clear
which tls interface is being tested.

Also nuke a few stray outdated comments in makefiles.

Finally, turn a TLS test on for arm32. This is supposed to work (the
other TLS interfaces have it on), but I'm unable to boot arm32 to verify.

Change-Id: I5dc086b26cbf79b18a1b7c6eb0ac6b39f22e8485