[musl] Uniformly stub out process, group, and user IDs
Note that Linux maintains some of these as a per-thread property, rather
than per-process. The setxid.c file once contained a very complicated
mechanism for attempting to emulate these as per-process on Linux. The
comments this patch removes in posix_spawn.c are referring to the same
thing.

These stubs, and presumably any more realistic mechanism we might choose
to invent, will supply these per-process, and there is no need for us to
have that weird indirection.

Change-Id: I2b810bdd8bd8ff8ab77b072c6b658ff5cf283ccc