[launchpad] New library for process startup and ELF loading
This adds a wholly new library called launchpad.  It will
eventually replace all the process startup and ELF loading code
in mxio.  For now it's only used by the new "launch" program,
which is a somewhat flexible test harness for manually testing
the launchpad interface.  The library does not yet have various
convenience functions for starting a process in standard ways.

Change-Id: I15ea77d83a65151cab8a18382f55dabf88cd5ec0