[syscalls] mx_debug_write should return number of chars written
When the caller wants to write more chars than the syscall allows
the syscall should return the number of chars it actually wrote
and not ERR_TOO_BIG so the caller can make additional calls for
the rest of the buffer.

This allows long printf()s from tests and the like to show up
when running in the non-graphical console, for example when
running in qemu without -g.

Change-Id: I77ceb826890acef807330408377b91f27b43590a