[ulib][gfx] integrate draw character support w/ gfx_putchar()
Less overhead than the old font code that called putpixel
on each pixel of the character being drawn.

For gfxconsole repainting kilo on skylake NUC, this cut
repaints from ~11.9ms to ~5.2ms.  That time does include
the off-to-on blits as well, the cost of which has not
changed.

An experimental hack to hold invalidates until after
the vt100 engine processes all characters took it down
to ~4ms.

Change-Id: I214e6b597205d71a0b96735e6cb5490e1ae79da8