[musl] Simplify dynamic linker startup
With some more clever linker tricks we can avoid the need for any
assembly code to let the dynamic linker discover its own base and
PT_DYNAMIC locations at runtime.  This also assumes that the system
starts up with the stack properly aligned for ABI requirements, so
there is no need for the dynamic linker to align the stack.

Change-Id: Iaa8e2699155521916e5c05b5edbf95959ce38e9a