Define errors in global header with codegen from definition file
This defines the error values shared across kernel and userspace in a
new header file in global/include/global/fuchsia-types.h derived from
a definition file called fuchsia-types.def that contains some
machinery to allow for further automated manipulation down the road,
for example spelling the symbols differently in different contexts or
providing mechanical strerror() routines.  The spelling and definition
of the ERR_FOO symbols remains the same in this patch (preprocessor
defines for integer constants).

Since the purpose of this header is to be shared across kernel and
userspace it is placed into a new top-level directory called 'global'
that is a peer of kernel/ and system/. This directory should contain
very few things.

This is only a first step to reduce duplication. Planned improvements
include defining status_t types for kernel user and mx_status_t types
for userspace, adding strerror() functions for the types, etc. The
expectation is that when we write a code generator for the syscall
dispatch that generator will be responsible for producing the correct
error definitions as well.

Change-Id: I5599a168e49def9f9555f081583a75bb02c61699