[kernel] Fix c++ allocations
c++ 'new' style allocations per-standard never fail, even
with the 'no-exceptions' flag passed to them. Therefore
the compiler elides checks of the style

   auto foo = new Foo(..);
   if (!foo) {
      return error;
   }

This is true even if |foo| is a smart pointer as long
as the compiler can see into the smart class.

bug MG-210 has different approches. This is the
one that seems more workable, other than obscure compiler
flags such as -fcheck-new.

In this CL we remove the standar new and new [] functions
and introduce overloads that take an AllocCheck* object
which enforces at runtime that check() is called.

Change-Id: Ie1b80c1ed4013929ebb2540ae7576ef5c08101f2