[uapp] [utest] Switch most uapp and utest builds to dynamic linking
Nearly every uapp and utest that can be dynamically linked is now.
The mxr-tls test is omitted because it's broken when dynamically
linked; that will probably be moot after the forthcoming TLS revamp.

A few test helper programs cannot use dynamic linking as written
because they get run with a nonstandard bootstrap message that does
not include the magic required for the dynamic linker's startup.

I didn't touch anything using ulib/ddk because I didn't bother to
figure out how that fits together.

Change-Id: I93b08960d946e23cd4ac89216e6d32789f99405a