[musl] Fix GNU-ish declaration of basename in string.h
This was originally declared as

    char* basename();

and incorrectly modified to

    char* basename(void);

when -Werror=strict-prototypes was introduced. musl declares it as such
presumably because glibc declares it complicatedly as either trafficking
in char* or in const char* depending on various things. Given that the
only implementation we provide in fact modifies the string, just declare
char* and call it a day.

Change-Id: Ia663f348099f01c238a9ca3690cb49624a1b12c3