[ulib] [musl] [ldso] Log via MXIO_LOGGER if available, fall back to mx_debug_write
This makes the dynamic linker check its bootstrap message for a
MXIO_LOGGER handle, and use mx_log_write on that handle for its
debug/error logging if possible, falling back to using the loader
service debug print RPC (as before) if possible, or to mx_debug_write if
all else fails.  Lack of a loader service and numerous other error
conditions now try to log a message and fail as gracefully as possible,
rather than intentionally crashing immediately.

Change-Id: If874e70f4e68a2426ae2f86fdac40d6bb750875f