[system][launchpad] Rename convenience functions.
The current naming could be improved.

The term "launchpad_launch" implies a general wrapper around the
launchpad API. Great. However, the name says nothing about mxio,
which is what launchpad_launch currently does.

Wrappers that include mxio support would be easier to read in code
if they had mxio in the name. And they would be less error prone to
use because one would know one is getting mxio support (e.g. stdio!).
So that is the high order bit of this change.

I have a need for passing additional handles to mxio subprocesses.
This patch adds another function that provides for this.
Another thought is to add the additional handle support to
launchpad_launch_mxio. Not sure which will be the common case,
and we can revisit this. But again the high order bit is to get "mxio"
into the names of wrappers that provide mxio support.

Thus:
launchpad_launch_basic -> launchpad_launch.
launchpad_launch -> launchpad_launch_mxio
launchpad_launch_mxio_etc: new

Change-Id: Iace907a433b5056bd0366fd8bc5b9d646dc5a8f3