[utils] Refactor traits for hashtables.
Split HashTraits so that all of the following are independently specified.

++ KeyType
++ PtrType
++ BucketType
++ HashType
++ NumBuckets
++ KeyTraits
++ HashTraits

Defaults have been provided for the key and hash traits which allows objects to
be stored in a hash table simply by defining an instanced GetKey method and
static GetHash method in the object to be stored.

The default implementation of HashTraits will automatically perform a mod
operation on the hash value returned by the objects GetHash method to ensure
that it is automatically in the range from [0, NumBuckets].  Users who can
already guarantee that this is the case (making the mod operation a waste of
time) can avoid this behavior by defining their own HashTraits which skips the
mod operation.

Bug: MG-230
Change-Id: Icf8ff1b8a5656470eb3823ae4fc9caacdf517d80