[kernel][magenta] IO Port refactor (part 2)
This change removes the fixed capacity of the IO Port. Instead
of a circular buffer, internally is a linked list of Packets
only limited by kernel memory.

The packets then can be of variable size, so the constrain
of either queueing mx_user_packet_t or mx_io_packet is removed
and mx_user_packet_t is also removed from public headers.

Change-Id: Ia418c31cf16eaada8d4b42d727335b33ab9aba76