[kernel] [userboot] Implement vDSO and new userboot scheme
libmagenta.so is now the vDSO, magically embedded in the kernel
and provided to userland as a read-only VMO at boot time.

The old statically linked userboot program is replaced by a new
userboot that is a pure read-only/reentrant DSO with a simplified
layout (like the vDSO).  The kernel loads this and the vDSO with
simple hard-wired code, eliminating the need for a full ELF loader
in the kernel.  The userboot image is embedded directly into the
kernel like the vDSO image, also eliminating the need for any
handling of the bootfs format in the kernel.

The kernel starts userboot with a message pipe handle as its
argument after writing a bootstrap message to the pipe.  This
message has the standard mx_proc_args_t format, using the kernel
command line as environment strings, and handles to the vDSO VMO and
the bootfs VMO.

userboot reads the bootfs to find the first real user program to
run.  It has a full ELF loader including PT_INTERP support, so this
program can use the dynamic linker, which is also the shared
libc.so; userboot does not implement the "loader service", so there
is no means to load other shared libraries from the bootfs and the
program must statically in link everything but libc itself.

userboot interprets the kernel command line for its own options
(so far just the "userboot=filename" option to change the default of
what to run next from "bin/devmgr").  It then passes on the
bootstrap message containing the kernel command line as the
environment strings and the VMO handles from the kernel, so the
startup of the first program is just like any other normal userland
program's startup and it can access (and propagate) the kernel
command line via its environment.

Change-Id: I7f11ee84aaf585de6544731b722a420c616a2a26