[ulib] [musl] Notice stack VMO handle and use it to compute stack bounds
The new protocol for process startup is that the creator (launchpad)
creates a VMO to back the stack, maps in that entire VMO, and starts
the process with the SP pointing at the end of it, while passing the
VMO handle in the startup message.  Thus the startup code can sample
the incoming SP to see where the end of the stack mapping is, query
the VMO to find its size, and thereby compute the bounds of the stack.

Change-Id: Ieaca5777e192429a348bf5ce957b1481c5007729