[gn] Process module list inside GN instead of generating files
This moves processing of the module list fully into GN generation and
moves the list of modules into a GN variable instead of relying on
pre-GN file generation. This makes each out directory contain the list
of modules configured for that out dir instead of relying on a
generated file in the source dir.

The way it works now is:
(gen.py time)
*) gen.py propagates the module list to a GN variable called 'modules'

(every time ninja runs)
*) //packages/gn:mkbootfs calls exec_script("process_modules.py") passing
in the list of modules and paths to the out dir being built
*) process_modules.py parses the module configs and writes the user
bootfs manifest to the gendir and emits a list of labels that the
bootfs depends on
*) all labels specified in the config are built, then the bootfs is
assembled

Change-Id: If873c36d80b33b373b2007cbe5bd54258e0a880f