[utils][containers] Add insert_or_find to containers.
Add an insert_or_find method to associative containers.  The method
will insert the supplied object into the collection if the object's
key is not already present in the collection, then indicate success by
returning true.  If a key collision occurs, the object will not be
added to the collection (and the pointer in the caller's scope will
remain unmodified, even if ::move'ed into the collection) and the
method will return false.

An optional pointer to an [out] iterator may be supplied with the
call.  In the case of a successful insert operation, the iterator will
point to the newly inserted member.  In the case of a failed insert
opertaion, the the iterator will point to the the object in the
collection whose key the new object collided with.

Also; change the API across collections to have find operations return
iterators to found elements instead of returning const PtrType
references.

Change-Id: Iad094bd6dec13aa8a1d0bba47dacb5a940ebb2ff