Add ftl::MakeRunnable
This function makes it easier to use lambdas with mojo::Callbacks.
Ideally we'd teach mojo::Callback to hold callbacks directly, but this
approach is easier in the near term.

Also, rename ftl::WrapLambda to ftl::MakeCopyable. The old name didn't
explain why you'd want to wrap a lambda. The new name says what the
function does.

Change-Id: I598fc8ca0a11ccc92da1668ebac303dd58ade18c