[ddk][usb] First cut at new USB stack
The existing coreboot based stack is replaced with a brand new XHCI driver and hub driver
To enable, build with ENABLE_NEW_USB=1
Once the old stack is removed and this is enabled by default,
work will continue on cleaning up the DDK USB protocols and APIs.

At this point the stack is basically functional.
However the device disconnect path is pretty sketchy.
More cleanup work in that area is needed.
More work is also needed for supporting SuperSpeed hubs

Change-Id: I97ae1678e503bf643139fa1b28666b74154f113a